pipelines: return a crawl_result object instead of just the record
This commit introduces the model CrawlResult, which wrap the record crawled and eventual errors occurred during the processing,
the remote file and the original source. Thic object is returned when `process_item` is called in the pipeline, instead of
the either the record crawled or the error traceback.

This commit also changes some unit and functional tests to adapt them to the new api.

* INCOMPATIBLE: this breaks the current API by returning a serialized CrawlResult instead of a record.

Signed-off-by: Antonio Cesarano <cesarano2607@gmail.com>